refactor(`Result`): Change the folder to utils
